Clean up the release process a bit
Include the full release directory prefix in the RELEASE-DIR variable.
This makes the makefile a bit less verbose.

Add release-dir target that creates the release directory and make use
of it. This decreases clutter in the console output since the release
directory is now created only once with `mkdir -p`

Remove some commented-out code.

Issue #35